Refactor flow
