+++ v2rayN/ServiceLib/Handler/Fmt/BaseFmt.cs
@@ -21,12 +21,6 @@ public class BaseFmt
 
     protected static int ToUriQuery(ProfileItem item, string? securityDef, ref Dictionary<string, string> dicQuery)
     {
-        var protocolExtra = item.GetProtocolExtra();
-        if (!protocolExtra.Flow.IsNullOrEmpty())
-        {
-            dicQuery.Add("flow", protocolExtra.Flow);
-        }
-
         if (item.StreamSecurity.IsNotEmpty())
         {
             dicQuery.Add("security", item.StreamSecurity);
@@ -209,9 +203,6 @@ public class BaseFmt
 
     protected static int ResolveUriQuery(NameValueCollection query, ref ProfileItem item)
     {
-        var protocolExtra = item.GetProtocolExtra();
-        protocolExtra.Flow = GetQueryValue(query, "flow");
-        item.SetProtocolExtra(protocolExtra);
         item.StreamSecurity = GetQueryValue(query, "security");
         item.Sni = GetQueryValue(query, "sni");
         item.Alpn = GetQueryDecoded(query, "alpn");
