+++ v2rayN/ServiceLib/Handler/Fmt/TrojanFmt.cs
@@ -10,6 +10,7 @@ public class TrojanFmt : BaseFmt
         {
             ConfigType = EConfigType.Trojan
         };
+        var protocolExtra = item.GetProtocolExtra();
 
         var url = Utils.TryUri(str);
         if (url == null)
@@ -23,8 +24,10 @@ public class TrojanFmt : BaseFmt
         item.Password = Utils.UrlDecode(url.UserInfo);
 
         var query = Utils.ParseQueryString(url.Query);
+        protocolExtra.Flow = GetQueryValue(query, "flow");
         ResolveUriQuery(query, ref item);
 
+        item.SetProtocolExtra(protocolExtra);
         return item;
     }
 
@@ -34,12 +37,17 @@ public class TrojanFmt : BaseFmt
         {
             return null;
         }
+        var protocolExtra = item.GetProtocolExtra();
         var remark = string.Empty;
         if (item.Remarks.IsNotEmpty())
         {
             remark = "#" + Utils.UrlEncode(item.Remarks);
         }
         var dicQuery = new Dictionary<string, string>();
+        if (!protocolExtra.Flow.IsNullOrEmpty())
+        {
+            dicQuery.Add("flow", protocolExtra.Flow);
+        }
         ToUriQuery(item, null, ref dicQuery);
 
         return ToUri(EConfigType.Trojan, item.Address, item.Port, item.Password, dicQuery, remark);
