+++ v2rayN/ServiceLib/Manager/ActionPrecheckManager.cs
@@ -1,5 +1,3 @@
-using ServiceLib.Common;
-
 namespace ServiceLib.Manager;
 
 /// <summary>
