+++ v2rayN/ServiceLib/Models/ProfileItem.cs
@@ -1,5 +1,3 @@
-using ServiceLib.Common;
-
 namespace ServiceLib.Models;
 
 [Serializable]
