+++ v2rayN/ServiceLib/Services/CoreConfig/Singbox/SingboxOutboundService.cs
@@ -1,5 +1,3 @@
-using System.Linq;
-
 namespace ServiceLib.Services.CoreConfig;
 
 public partial class CoreConfigSingboxService
