+++ v2rayN/ServiceLib/Services/CoreConfig/V2ray/V2rayOutboundService.cs
@@ -1,5 +1,3 @@
-using System.Linq;
-
 namespace ServiceLib.Services.CoreConfig;
 
 public partial class CoreConfigV2rayService
